# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dogly API",
    "description": "Public API V1 (Spec V1 sez. 9). OpenAPI is the mobile client contract.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/digestive/fecal/init": {
      "post": {
        "tags": [
          "digestive"
        ],
        "summary": "Init Fecal",
        "operationId": "init_fecal_v1_digestive_fecal_init_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FecalInitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FecalInitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/digestive/fecal/{event_id}/complete": {
      "post": {
        "tags": [
          "digestive"
        ],
        "summary": "Complete Fecal",
        "operationId": "complete_fecal_v1_digestive_fecal__event_id__complete_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FecalCompleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/digestive/events/{event_id}": {
      "get": {
        "tags": [
          "digestive"
        ],
        "summary": "Get Digestive Event",
        "operationId": "get_digestive_event_v1_digestive_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigestiveEventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/digestive/events/{event_id}/context": {
      "patch": {
        "tags": [
          "digestive"
        ],
        "summary": "Update Digestive Context",
        "description": "Save one sparse owner answer and deterministically refresh the result.",
        "operationId": "update_digestive_context_v1_digestive_events__event_id__context_patch",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DigestiveContextUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigestiveEventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/dogs/{dog_id}/digestive-summary": {
      "get": {
        "tags": [
          "digestive"
        ],
        "summary": "Get Digestive Summary",
        "operationId": "get_digestive_summary_v1_dogs__dog_id__digestive_summary_get",
        "parameters": [
          {
            "name": "dog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dog Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigestiveSummaryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfidenceBand": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH"
        ],
        "title": "ConfidenceBand"
      },
...
      "DigestiveContextUpdateRequest": {
        "properties": {
          "vomiting_today": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vomiting Today"
          },
          "reduced_activity_today": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reduced Activity Today"
          },
          "unusual_food_48h": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unusual Food 48H"
          }
        },
        "type": "object",
        "title": "DigestiveContextUpdateRequest",
        "description": "Sparse owner-confirmed context; omitted answers remain unknown."
      },
      "DigestiveEventOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "dog_id": {
            "type": "string",
            "title": "Dog Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "fecal_score_estimate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecal Score Estimate"
          },
          "consistency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consistency"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "image_quality": {
            "type": "string",
            "enum": [
              "sufficient",
              "insufficient",
              "unknown"
            ],
            "title": "Image Quality",
            "default": "unknown"
          },
          "quality_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Quality Warnings"
          },
          "mucus_candidate": {
            "type": "string",
            "title": "Mucus Candidate",
            "default": "unknown"
          },
          "fresh_blood_candidate": {
            "type": "string",
            "title": "Fresh Blood Candidate",
            "default": "unknown"
          },
          "melena_candidate": {
            "type": "string",
            "title": "Melena Candidate",
            "default": "unknown"
          },
          "foreign_material_candidate": {
            "type": "string",
            "title": "Foreign Material Candidate",
            "default": "unknown"
          },
          "confidence_band": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceBand"
              },
              {
                "type": "null"
              }
            ]
          },
          "safety_flags": {
            "items": {
              "$ref": "#/components/schemas/SafetyFlag"
            },
            "type": "array",
            "title": "Safety Flags"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "active_food_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Food Name"
          },
          "baseline_comparison": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "INSUFFICIENT",
                  "BELOW_USUAL",
                  "NEAR_USUAL",
                  "ABOVE_USUAL"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Comparison"
          },
          "intelligence_schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intelligence Schema Version"
          },
          "overall_state": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ROUTINE",
                  "MONITOR",
                  "ATTENTION",
                  "VET_CONTACT"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall State"
          },
          "consumer_headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consumer Headline"
          },
          "consumer_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consumer Summary"
          },
          "relevant_context": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Relevant Context"
          },
          "possible_associations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Possible Associations"
          },
          "safety_state": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ROUTINE",
                  "MONITOR",
                  "ATTENTION",
                  "VET_CONTACT"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Safety State"
          },
          "recommended_next_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Next Step"
          },
          "followup_key": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "vomiting_today",
                  "reduced_activity_today",
                  "unusual_food_48h"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Followup Key"
          },
          "followup_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Followup Question"
          },
          "what_to_watch": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "What To Watch"
          },
          "observation_reliability": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observation Reliability"
          },
          "knowledge_references": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Knowledge References"
          },
          "reasoning_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Version"
          },
          "baseline_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dog_id",
          "status",
          "created_at"
        ],
        "title": "DigestiveEventOut"
      },
      "DigestiveSummaryOut": {
        "properties": {
          "dog_id": {
            "type": "string",
            "title": "Dog Id"
          },
          "rolling_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rolling Score"
          },
          "variability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variability"
          },
          "data_sufficiency": {
            "type": "string",
            "title": "Data Sufficiency",
            "default": "insufficient"
          },
          "recent_trend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recent Trend"
          },
          "safety_flags": {
            "items": {
              "$ref": "#/components/schemas/SafetyFlag"
            },
            "type": "array",
            "title": "Safety Flags"
          }
        },
        "type": "object",
        "required": [
          "dog_id"
        ],
        "title": "DigestiveSummaryOut"
      },
...
      "FecalCompleteResponse": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "status"
        ],
        "title": "FecalCompleteResponse"
      },
      "FecalInitRequest": {
        "properties": {
          "dog_id": {
            "type": "string",
            "title": "Dog Id"
          },
          "client_request_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Client Request Id"
          },
          "bytes": {
            "type": "integer",
            "maximum": 12000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Bytes"
          },
          "content_type": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ],
            "title": "Content Type",
            "default": "image/jpeg"
          }
        },
        "type": "object",
        "required": [
          "dog_id",
          "client_request_id",
          "bytes"
        ],
        "title": "FecalInitRequest"
      },
      "FecalInitResponse": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "upload": {
            "$ref": "#/components/schemas/SignedUpload"
          },
          "quota_reserved": {
            "type": "boolean",
            "title": "Quota Reserved",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "status",
          "upload"
        ],
        "title": "FecalInitResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SafetyFlag": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "default": "info"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code"
        ],
        "title": "SafetyFlag",
        "description": "Structured flags consumed by the deterministic copy layer (sez. 16.3).\nGenerated text may never downgrade a safety flag (sez. 19.3)."
      },
...
      "SignedUpload": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "storage_path": {
            "type": "string",
            "title": "Storage Path"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "url",
          "storage_path",
          "expires_at"
        ],
        "title": "SignedUpload"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}